moved directory locations and updated build
