+++ .replit
@@ -1,3 +1,63 @@
 language = "java10"
 #run = "echo java RunEdgeConvert"
 run = "sh build.sh"
+entrypoint = "Main.java"
+hidden = ["**/*.class"]
+compile = "javac -classpath .:target/dependency/* -d . $(find . -type f -name '*.java')"
+
+[languages]
+
+[languages.java]
+pattern = "**/*.java"
+
+[languages.java.languageServer]
+start = "jdt-language-server"
+
+[debugger]
+support = true
+
+[debugger.compile]
+command = "javac -classpath .:/run_dir/junit-4.12.jar:target/dependency/* -g -d . $(find . -type f -name '*.java')"
+
+[debugger.interactive]
+transport = "localhost:0"
+connectTimeout = 60
+startCommand = "java-debug"
+
+[debugger.interactive.initializeMessage]
+command = "initialize"
+type = "request"
+
+[debugger.interactive.initializeMessage.arguments]
+adapterID = "cppdbg"
+clientID = "replit"
+clientName = "replit.com"
+columnsStartAt1 = true
+linesStartAt1 = true
+locale = "en-us"
+pathFormat = "path"
+supportsInvalidatedEvent = true
+supportsProgressReporting = true
+supportsRunInTerminalRequest = true
+supportsVariablePaging = true
+supportsVariableType = true
+
+[debugger.interactive.launchMessage]
+command = "launch"
+type = "request"
+
+[debugger.interactive.launchMessage.arguments]
+classPaths = ["."]
+mainClass = "Main"
+
+[packager]
+language = "java"
+
+[packager.features]
+packageSearch = true
+
+[unitTest]
+language = "java"
+
+[nix]
+channel = "stable-22_11"
