+++ replit.nix
@@ -1,0 +1,8 @@
+{ pkgs }: {
+    deps = [
+        pkgs.graalvm17-ce
+        pkgs.maven
+        pkgs.replitPackages.jdt-language-server
+        pkgs.replitPackages.java-debug
+    ];
+}
