+++ src/main/java/EdgeConnector.java
